utility/loader/pyproject.toml
[project]
name = "loader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "azure-identity>=1.25.1",
    "azure-search-documents>=11.6.0",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
